chore(issue-filter): add ability to create webhook filter
